Share interface
